backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pymysql==1.1.0
python-dotenv==1.0.1
typer==0.12.5
click==8.1.7
rich==13.7.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
# Dev dependencies
ruff==0.1.15
pytest==8.0.0
httpx==0.26.0
